+++ api/.eslintrc.js
@@ -11,6 +11,7 @@ module.exports = {
   },
   rules: {
     "prettier/prettier": ["warn"],
+    "no-unused-vars": "warn",
   },
   plugins: ["jest", "prettier"],
 };